Remove dothash support
This update removes dothash. Although it works correctly in Spark's tests, it (for whatever reason) doesn't work well in our marketing site. I also can't seem to break it in Spark's unit tests.

To allow for our marketing site to use the latest version of Spark, we'll need to remove dothash.

This version needs to be minor bumped.